# api/openapi.yaml
openapi: 3.0.3
info:
  title: Shepherd API
  description: Background coding agent orchestrator on Kubernetes.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
...
  /api/v1/tasks:
    post:
      operationId: createTask
      summary: Create a new agent task
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Compressed context exceeds size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Content-Type must be application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: listTasks
      summary: List agent tasks
      tags: [tasks]
      parameters:
        - name: repo
          in: query
          description: Filter by shepherd.io/repo label
          schema:
            type: string
        - name: issue
          in: query
          description: Filter by shepherd.io/issue label
          schema:
            type: string
        - name: fleet
          in: query
          description: Filter by shepherd.io/fleet label
          schema:
            type: string
        - name: active
          in: query
          description: If "true", only return non-terminal tasks
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
  /api/v1/tasks/{taskID}:
    get:
      operationId: getTask
      summary: Get a single task
      tags: [tasks]
      parameters:
        - $ref: "#/components/parameters/taskID"
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/tasks/{taskID}/status:
    post:
      operationId: updateTaskStatus
      summary: Update task status (runner callback)
      tags: [internal]
      parameters:
        - $ref: "#/components/parameters/taskID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdateRequest"
      responses:
        "200":
          description: Status accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusAcceptedResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Content-Type must be application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    taskID:
      name: taskID
      in: path
      required: true
      schema:
        type: string
  schemas:
    CreateTaskRequest:
      type: object
      required: [repo, task, callbackURL, runner]
      properties:
        repo:
          $ref: "#/components/schemas/RepoRequest"
        task:
          $ref: "#/components/schemas/TaskRequest"
        callbackURL:
          type: string
          format: uri
        runner:
          $ref: "#/components/schemas/RunnerConfig"
        labels:
          type: object
          additionalProperties:
            type: string
    RepoRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        ref:
          type: string
    TaskRequest:
      type: object
      required: [description]
      properties:
        description:
          type: string
        context:
          type: string
        sourceURL:
          type: string
        sourceType:
          type: string
        sourceID:
          type: string
    RunnerConfig:
      type: object
      required: [sandboxTemplateName]
      properties:
        sandboxTemplateName:
          type: string
        timeout:
          type: string
        serviceAccountName:
          type: string
    TaskResponse:
      type: object
      required: [id, namespace, repo, task, callbackURL, status, createdAt]
      properties:
        id:
          type: string
        namespace:
          type: string
        repo:
          $ref: "#/components/schemas/RepoRequest"
        task:
          $ref: "#/components/schemas/TaskRequest"
        callbackURL:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatusSummary"
        createdAt:
          type: string
          format: date-time
        completionTime:
          type: string
          format: date-time
          nullable: true
    TaskStatusSummary:
      type: object
      required: [phase, message]
      properties:
        phase:
          type: string
        message:
          type: string
        sandboxClaimName:
          type: string
        prURL:
          type: string
        error:
          type: string
    StatusUpdateRequest:
      type: object
      required: [event]
      properties:
        event:
          type: string
          enum: [started, progress, completed, failed]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    StatusAcceptedResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
        note:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string